collect_articles/requirements.txt
# Основные библиотеки для парсинга и HTTP‑запросов
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0

# Работа с PostgreSQL
psycopg2-binary>=2.9.0

# Машинное обучение и эмбеддинги
torch>=2.0.0  # PyTorch для генерации эмбеддингов
transformers>=4.30.0  # Hugging Face Transformers
sentence-transformers>=2.2.0  # Готовые модели для эмбеддингов
numpy>=1.24.0
scipy>=1.10.0

# Утилиты и вспомогательные библиотеки
python-dotenv>=1.0.0  # Загрузка переменных окружения из .env
tqdm>=4.65.0  # Индикаторы прогресса
python-dateutil>=2.8.0  # Работа с датами

# Логирование
loguru>=0.7.0  # Удобное логирование

# Веб‑скрапинг и асинхронные запросы (опционально)
aiohttp>=3.8.0  # Асинхронные HTTP‑запросы
selenium>=4.0.0  # Для сложных случаев парсинга (JS‑рендеринг)

# Тестирование (опционально)
pytest>=7.0.0  # Фреймворк для тестирования
pytest-asyncio>=0.21.0  # Поддержка async в тестах
